Reorganize application setup process
Move setupTempDir from cli.ts to main.ts and integrate it into initialiseApplication.
Rename initialiseOptionsFromCommandLine to parseArgsToAppArguments for clarity.
Move logging initialization to app.ts for more explicit control.
This improves separation of concerns and makes side-effects more explicit.
