+++ test/app/server-tests.ts
@@ -44,6 +44,12 @@ function createMockAppArgs(overrides: Partial<AppArguments> = {}): AppArguments 
         useLocalProps: true,
         propDebug: false,
         tmpDir: undefined,
+        isWsl: false,
+        loggingOptions: {
+            debug: false,
+            suppressConsoleLog: false,
+            paperTrailIdentifier: 'test',
+        },
         ...overrides,
     };
 }
